Crate iota_kvstore

Modules

client

Implementation of key-value store reader and writer traits for the BigTable client.

proto

Compiled protobuf definitions.

Structs

BigTableClient

A high-level client for interacting with Google BigTable using authenticated requests over gRPC.

Cell

Represents a single cell in a BigTable row.

Checkpoint

Represents all stored Key-Value data associated to a checkpoint containing both the summary and the full contents.

KvWorker

This worker implementation is responsible for processing checkpoints by storing its data as Key-Value pairs. The Key-Value pairs are stored in a BigTableDB.

Row

Represents a row in a BigTable table.

TransactionData

Represents all stored Key-Value data associated with a transaction, including its effects, events, and the checkpoint number it belongs to.

Traits

KeyValueStoreReader

Read key-value data from a persistent store, such as objects, transactions, and checkpoints.

KeyValueStoreWriter

Writing key-value data to a persistent store, such as objects, transactions, and checkpoints.