iota_kvstore

Function transactions_by_address

pub fn transactions_by_address<'a>(
    checkpoint: &'a CheckpointData,
) -> impl Iterator<Item = (Address, TransactionSequenceNumber, TransactionDigest)> + 'a

Uses CheckpointData to map addresses affected by a transaction to the respective transaction identifiers.

Affected addresses include the sender, and payer of the transaction, plus the owners of all changed objects.