iota_kvstore::proto::bigtable

Module v2

Modules

bigtable_client

Generated client implementations.

column_range

Nested message and enum types in ColumnRange.

execute_query_request

Nested message and enum types in ExecuteQueryRequest.

execute_query_response

Nested message and enum types in ExecuteQueryResponse.

mutate_rows_request

Nested message and enum types in MutateRowsRequest.

mutate_rows_response

Nested message and enum types in MutateRowsResponse.

mutation

Nested message and enum types in Mutation.

partial_result_set

Nested message and enum types in PartialResultSet.

prepare_query_request

Nested message and enum types in PrepareQueryRequest.

read_change_stream_request

Nested message and enum types in ReadChangeStreamRequest.

read_change_stream_response

Nested message and enum types in ReadChangeStreamResponse.

read_modify_write_rule

Nested message and enum types in ReadModifyWriteRule.

read_rows_request

Nested message and enum types in ReadRowsRequest.

read_rows_response

Nested message and enum types in ReadRowsResponse.

request_stats

Nested message and enum types in RequestStats.

result_set_metadata

Nested message and enum types in ResultSetMetadata.

row_filter

Nested message and enum types in RowFilter.

row_range

Nested message and enum types in RowRange.

type

Nested message and enum types in Type.

value

Nested message and enum types in Value.

value_range

Nested message and enum types in ValueRange.

Structs

ArrayValue

ArrayValue is an ordered list of Value.

Cell

Specifies (some of) the contents of a single row/column/timestamp of a table.

CheckAndMutateRowRequest

Request message for Bigtable.CheckAndMutateRow.

CheckAndMutateRowResponse

Response message for Bigtable.CheckAndMutateRow.

Column

Specifies (some of) the contents of a single row/column intersection of a table.

ColumnMetadata

Describes a column in a Bigtable Query Language result set.

ColumnRange

Specifies a contiguous range of columns within a single column family. The range spans from <column_family>:<start_qualifier> to <column_family>:<end_qualifier>, where both bounds can be either inclusive or exclusive.

ExecuteQueryRequest

Request message for Bigtable.ExecuteQuery

ExecuteQueryResponse

Response message for Bigtable.ExecuteQuery

Family

Specifies (some of) the contents of a single row/column family intersection of a table.

FullReadStatsView

FullReadStatsView captures all known information about a read.

GenerateInitialChangeStreamPartitionsRequest

NOTE: This API is intended to be used by Apache Beam BigtableIO. Request message for Bigtable.GenerateInitialChangeStreamPartitions.

GenerateInitialChangeStreamPartitionsResponse

NOTE: This API is intended to be used by Apache Beam BigtableIO. Response message for Bigtable.GenerateInitialChangeStreamPartitions.

Idempotency

Parameters on mutations where clients want to ensure idempotency (i.e. at-most-once semantics). This is currently only needed for certain aggregate types.

MutateRowRequest

Request message for Bigtable.MutateRow.

MutateRowResponse

Response message for Bigtable.MutateRow.

MutateRowsRequest

Request message for BigtableService.MutateRows.

MutateRowsResponse

Response message for BigtableService.MutateRows.

Mutation

Specifies a particular change to be made to the contents of a row.

PartialResultSet

A partial result set from the streaming query API. Cloud Bigtable clients buffer partial results received in this message until a resume_token is received.

PingAndWarmRequest

Request message for client connection keep-alive and warming.

PingAndWarmResponse

Response message for Bigtable.PingAndWarm connection keepalive and warming.

PrepareQueryRequest

Request message for Bigtable.PrepareQuery

PrepareQueryResponse

Response message for Bigtable.PrepareQueryResponse

ProtoFormat

Protocol buffers format descriptor, as described by Messages ProtoSchema and ProtoRows

ProtoRows

Rows represented in proto format.

ProtoRowsBatch

A part of a serialized ProtoRows message.

ProtoSchema

ResultSet schema in proto format

RateLimitInfo

Information about how client should adjust the load to Bigtable.

ReadChangeStreamRequest

NOTE: This API is intended to be used by Apache Beam BigtableIO. Request message for Bigtable.ReadChangeStream.

ReadChangeStreamResponse

NOTE: This API is intended to be used by Apache Beam BigtableIO. Response message for Bigtable.ReadChangeStream.

ReadIterationStats

ReadIterationStats captures information about the iteration of rows or cells over the course of a read, e.g. how many results were scanned in a read operation versus the results returned.

ReadModifyWriteRowRequest

Request message for Bigtable.ReadModifyWriteRow.

ReadModifyWriteRowResponse

Response message for Bigtable.ReadModifyWriteRow.

ReadModifyWriteRule

Specifies an atomic read/modify/write operation on the latest value of the specified column.

ReadRowsRequest

Request message for Bigtable.ReadRows.

ReadRowsResponse

Response message for Bigtable.ReadRows.

RequestLatencyStats

RequestLatencyStats provides a measurement of the latency of the request as it interacts with different systems over its lifetime, e.g. how long the request took to execute within a frontend server.

RequestStats

RequestStats is the container for additional information pertaining to a single request, helpful for evaluating the performance of the sent request. Currently, there are the following supported methods:

ResultSetMetadata

Describes the structure of a Bigtable result set.

Row

Specifies the complete (requested) contents of a single row of a table. Rows which exceed 256MiB in size cannot be read in full.

RowFilter

Takes a row as input and produces an alternate view of the row based on specified rules. For example, a RowFilter might trim down a row to include just the cells from columns matching a given regular expression, or might return all the cells of a row but not their values. More complicated filters can be composed out of these components to express requests such as, “within every column of a particular family, give just the two most recent cells which are older than timestamp X.”

RowRange

Specifies a contiguous range of rows.

RowSet

Specifies a non-contiguous set of rows.

SampleRowKeysRequest

Request message for Bigtable.SampleRowKeys.

SampleRowKeysResponse

Response message for Bigtable.SampleRowKeys.

StreamContinuationToken

NOTE: This API is intended to be used by Apache Beam BigtableIO. The information required to continue reading the data from a StreamPartition from where a previous read left off.

StreamContinuationTokens

NOTE: This API is intended to be used by Apache Beam BigtableIO. The information required to continue reading the data from multiple StreamPartitions from where a previous read left off.

StreamPartition

NOTE: This API is intended to be used by Apache Beam BigtableIO. A partition of a change stream.

TimestampRange

Specified a contiguous range of microsecond timestamps.

Type

Type represents the type of data that is written to, read from, or stored in Bigtable. It is heavily based on the GoogleSQL standard to help maintain familiarity and consistency across products and features.

Value

Value represents a dynamically typed value. The typed fields in Value are used as a transport encoding for the actual value (which may be of a more complex type). See the documentation of the Type message for more details.

ValueRange

Specifies a contiguous range of raw byte values.