iota_kvstore

Trait KeyValueStoreReader

pub trait KeyValueStoreReader {
    type Error;

    // Required methods
    fn get_objects<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        objects: &'life1 [ObjectKey],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Object>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_transactions<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        transactions: &'life1 [TransactionDigest],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionData>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_transaction_digests_by_address<'life0, 'async_trait>(
        &'life0 mut self,
        address: Address,
        cursor: impl 'async_trait + Into<Option<TransactionSequenceNumber>> + Send,
        limit: impl 'async_trait + TryInto<NonZeroUsize> + Send,
        order: TransactionsOrder,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<(TransactionSequenceNumber, TransactionDigest)>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_checkpoints<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        sequence_numbers: &'life1 [CheckpointSequenceNumber],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Checkpoint>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_checkpoints_by_digest<'life0, 'async_trait, I>(
        &'life0 mut self,
        digests: I,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Checkpoint>, Self::Error>> + Send + 'async_trait>>
       where I: IntoIterator<Item = CheckpointDigest> + Send + 'async_trait,
             I::IntoIter: Send,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn get_checkpoint_sequence_numbers<'life0, 'async_trait, I>(
        &'life0 mut self,
        digests: I,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<CheckpointSequenceNumber>, Self::Error>> + Send + 'async_trait>>
       where I: IntoIterator<Item = CheckpointDigest> + Send + 'async_trait,
             I::IntoIter: Send,
             Self: 'async_trait,
             'life0: 'async_trait;
}

Read key-value data from a persistent store, such as objects, transactions, and checkpoints.

Required Associated Types

type Error

Required Methods

fn get_objects<'life0, 'life1, 'async_trait>( &'life0 mut self, objects: &'life1 [ObjectKey], ) -> Pin<Box<dyn Future<Output = Result<Vec<Object>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetches a list of objects by their keys.

Not found objects are omitted from the output list.

fn get_transactions<'life0, 'life1, 'async_trait>( &'life0 mut self, transactions: &'life1 [TransactionDigest], ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionData>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetches a list of transactions by their digests.

Not found transactions are omitted from the output list.

fn get_transaction_digests_by_address<'life0, 'async_trait>( &'life0 mut self, address: Address, cursor: impl 'async_trait + Into<Option<TransactionSequenceNumber>> + Send, limit: impl 'async_trait + TryInto<NonZeroUsize> + Send, order: TransactionsOrder, ) -> Pin<Box<dyn Future<Output = Result<Vec<(TransactionSequenceNumber, TransactionDigest)>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Fetches a list (sequence number, digest) pairs for transactions affecting the given address, ordered by TransactionsOrder and capped at limit.

Pagination

cursor is exclusive.

TransactionsOrder::NewestFirst: returns entries with tx_seq < cursor.

TransactionsOrder::OldestFirst: returns entries with tx_seq > cursor.

When None, the scan starts from the beginning of the requested TransactionsOrder.

fn get_checkpoints<'life0, 'life1, 'async_trait>( &'life0 mut self, sequence_numbers: &'life1 [CheckpointSequenceNumber], ) -> Pin<Box<dyn Future<Output = Result<Vec<Checkpoint>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetches a list of checkpoints by their sequence numbers.

Not found checkpoints are omitted from the output list.

fn get_checkpoints_by_digest<'life0, 'async_trait, I>( &'life0 mut self, digests: I, ) -> Pin<Box<dyn Future<Output = Result<Vec<Checkpoint>, Self::Error>> + Send + 'async_trait>>
where I: IntoIterator<Item = CheckpointDigest> + Send + 'async_trait, I::IntoIter: Send, Self: 'async_trait, 'life0: 'async_trait,

Fetches a list of checkpoints by their digests.

Not found checkpoints are omitted from the output list.

fn get_checkpoint_sequence_numbers<'life0, 'async_trait, I>( &'life0 mut self, digests: I, ) -> Pin<Box<dyn Future<Output = Result<Vec<CheckpointSequenceNumber>, Self::Error>> + Send + 'async_trait>>
where I: IntoIterator<Item = CheckpointDigest> + Send + 'async_trait, I::IntoIter: Send, Self: 'async_trait, 'life0: 'async_trait,

Fetches a list of checkpoint sequence numbers by their digests.

Not found checkpoints are omitted from the output list.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.