iota_kvstore

Trait KeyValueStoreReader

pub trait KeyValueStoreReader {
    type Error;

    // Required methods
    fn get_objects<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        objects: &'life1 [ObjectKey],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Object>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_transactions<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        transactions: &'life1 [TransactionDigest],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionData>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_checkpoints<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        sequence_numbers: &'life1 [CheckpointSequenceNumber],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Checkpoint>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_checkpoints_by_digest<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        digests: &'life1 [CheckpointDigest],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Checkpoint>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Read key-value data from a persistent store, such as objects, transactions, and checkpoints.

Required Associated Types

type Error

Required Methods

fn get_objects<'life0, 'life1, 'async_trait>( &'life0 mut self, objects: &'life1 [ObjectKey], ) -> Pin<Box<dyn Future<Output = Result<Vec<Object>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetches a list of objects by their keys.

Not found objects are omitted from the output list.

fn get_transactions<'life0, 'life1, 'async_trait>( &'life0 mut self, transactions: &'life1 [TransactionDigest], ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionData>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetches a list of transactions by their digests.

Not found transactions are omitted from the output list.

fn get_checkpoints<'life0, 'life1, 'async_trait>( &'life0 mut self, sequence_numbers: &'life1 [CheckpointSequenceNumber], ) -> Pin<Box<dyn Future<Output = Result<Vec<Checkpoint>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetches a list of checkpoints by their sequence numbers.

Not found checkpoints are omitted from the output list.

fn get_checkpoints_by_digest<'life0, 'life1, 'async_trait>( &'life0 mut self, digests: &'life1 [CheckpointDigest], ) -> Pin<Box<dyn Future<Output = Result<Vec<Checkpoint>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Fetches a list of checkpoints by their digests.

Not found checkpoints are omitted from the output list.